requirements_test.txt
pytest>=8.0.0
pytest-django>=4.8.0
playwright>=1.40.0
pytest-playwright>=0.4.0
python-decouple>=3.8
faker>=20.0.0
requests>=2.31.0
allure-pytest>=2.13.0
